mlnx_qos: Handle total port buffer size specially
DCB_ATTR_DCB_BUFFER query responds with a structure that includes total
port buffer size. Return it from get_ieee_dcb_buffer() as a separate field,
instead of it being the last member of the buffer size array. Show it in
pretty_print. Extend set_dcb_buffer() to take it as an extra argument.

Signed-off-by: Petr Machata <petrm@nvidia.com>