mlnx_qos: Support cable lengths of up to 0xffff
The payload of DCB_ATTR_IEEE_PFC attribute is struct ieee_pfc. One of the
fields is an u16 delay, which mlnx_qos configures through the cable_len
parameter. However it arbitrarily caps the value at 1000, while to DCB
delay values of full 16 bits are allowed by the Linux kernel. Therefore
permit the value to go all the way up to 0xffff.

Signed-off-by: Petr Machata <petrm@nvidia.com>